refactored zfind so that redmonster is an optional dependency
