+++ py/desispec/zfind/__init__.py
@@ -1,0 +1,1 @@
+from .zfind import ZfindBase
